requirements.txt
crewai
crewai_tools
streamlit
python-dotenv
requests